payment-service/requirements.txt
Flask
Flask-Cors
Flask-SQLAlchemy
PyJWT
requests
PyMySQL
gunicorn
cryptography